Create VNP_PMM_within_R
This file contains the Vague Normal Priors Pattern Mixture Model that addresses response bias effects on treatment effect on a binary outcome; the means of which is providing a somewhat vague or minimally informative interval specification to the hypothesized response rate difference between those that respond positively (Y=1) vs. those in general, in each intervention arm. The main focus of this work is specifying (d0m, d0M), and (d1m, d1M), these are the lower and upper limit pairs for the prior interval of response difference as discussed before, for each intervention arm labeled 0 and 1, respectively.